fix(app): adapt buildRecording and crash recovery to target-rate app temps
The capture session now writes the app temp as 16 kHz mono, so the app
module's assumptions about raw device-rate temps are updated:

- A temp already at the target rate is never re-rated by the mic-duration
  heuristic: if the mic dies mid-recording (restart retries exhausted), the
  shortened reference duration would make crossCheckAppRate infer a bogus
  higher rate and time-warp the healthy app track -- the exact bug class
  this branch eliminates. The cross-check still guards the fallback/legacy
  raw path.
- CaptureFormat now describes the expected FILE format (16 kHz mono), so the
  rate/channel mismatch warnings flag the unexpected raw-write fallback
  instead of firing on every clean recording.
- Crash recovery reads the orphaned temp as 16 kHz mono instead of the
  device capture format. (A temp written by a pre-upgrade version is
  mis-rated -- the headerless temp carries no format; narrow window, best
  effort.)
- downmixToMono delegates to the shared AudioTapLib implementation.

The crash-recovery tests move to DualSourceRecorderCrashRecoveryTests to
keep the test class under the type-body-length cap.

Refs #379
